Show recommendations in Client view

+++ BistroQ.Domain/Models/Entities/NutritionFact.cs
@@ -4,11 +4,11 @@ public class NutritionFact
 {
     public int ProductId { get; set; }
 
-    public double? Calories { get; set; }
-    public double? Fat { get; set; }
-    public double? Fiber { get; set; }
-    public double? Protein { get; set; }
-    public double? Carbohydrates { get; set; }
+    public decimal? Calories { get; set; }
+    public decimal? Fat { get; set; }
+    public decimal? Fiber { get; set; }
+    public decimal? Protein { get; set; }
+    public decimal? Carbohydrates { get; set; }
 
     public virtual Product Product { get; set; } = null!;
 }
